move WithHealth interface to combatactions
