+++ 2023-09-10-ts/src/combat-actions/index.ts
@@ -1,5 +1,9 @@
 export class InvalidTargetError extends Error {}
 export class InvalidSourceError extends Error {}
+export interface WithHealth {
+  health: number;
+  isAlive: boolean;
+}
 
 export { AttackAction } from "./attack-action";
 export { HealingAction } from "./healing-action";
